+++ plugins/11_TheWarWithin/CHANGELOG.md
@@ -1,3 +1,7 @@
+# v105
+
+* Add Arathi Prisoner locations for the achievement [Back to the Wall](https://www.wowhead.com/achievement=40620/back-to-the-wall)
+
 # v104
 
 * Localization updates: zhTW